One tessellation, shared by both regions a frame carries
`wl_region` has no notion of curves, so a rounded container reaches the
compositor as a union of axis-aligned rectangles. That translation lived in
`blur`, and it is not about blur: the surface's input region needs the same
rectangles from the same shapes, and a second copy of a tessellation is a
second place for a corner to be wrong.

So it moves to `region`, with its tests, and `RegionRect` is the currency both
halves speak. The whole-pixel rounding the platform had written out by hand for
a surface's declared rectangles moves with it and gains the sentence saying why
it rounds outward.
